code refactor - add class method `get_workers_progress`
